feat: renamed feature.module to myFeature.module
This was done to show that the developer has to think of a name for the feature module
