+++ src/app/myFeature/myFeature.module.ts
@@ -1,11 +1,11 @@
-import { routes } from './feature.routes';
+import { routes } from './myFeature.routes';
 import { CommonModule } from '@angular/common';
 import { StoreModule } from '@ngrx/store';
 import { RouterModule, Routes } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { LetModule } from '@ngrx/component';
 import { EntitySearchComponent } from './pages/entity-search/entity-search.component';
-import { featureFeature } from './feature.reducers';
+import { myFeatureFeature } from './myFeature.reducers';
 import { EffectsModule } from '@ngrx/effects';
 import { EntitySearchEffects } from './pages/entity-search/entity-search.effects';
 import { CoreModule } from '../core/core.module';
@@ -16,9 +16,9 @@ import { CoreModule } from '../core/core.module';
     CommonModule,
     LetModule,
     RouterModule.forChild(routes),
-    StoreModule.forFeature(featureFeature),
+    StoreModule.forFeature(myFeatureFeature),
     EffectsModule.forFeature([EntitySearchEffects]),
     CoreModule,
   ],
 })
-export class FeatureModule {}
+export class MyFeatureModule {}
