+++ src/app/myFeature/myFeature.routes.ts
