+++ src/app/myFeature/myFeature.state.ts
@@ -1,5 +1,5 @@
 import { EntitySearchState } from './pages/entity-search/entity-search.state';
 
-export interface FeatureState {
+export interface MyFeatureState {
   entitySearch: EntitySearchState;
 }
