+++ src/app/myFeature/pages/entity-search/entity-search.selectors.spec.ts
